fix: some type-fu

+++ waku/waku_rln_relay/group_manager/on_chain/group_manager.nim
@@ -856,7 +856,7 @@ method isReady*(g: OnchainGroupManager): Future[bool] {.async.} =
     cast[BlockNumber](await g.ethRpc.get().provider.eth_blockNumber())
 
   # the node is still able to process messages if it is behind the latest block by a factor of the valid roots
-  if g.latestProcessedBlock < currentBlock - g.validRoots.len:
+  if uint(g.latestProcessedBlock) < uint(currentBlock) - uint(g.validRoots.len):
     return false
 
   return not (await g.isSyncing())
